requirements.txt
absl-py==2.3.1
accelerate==1.9.0
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
diffusers==0.34.0
distro==1.9.0
einops==0.8.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.5
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
joblib==1.5.1
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
openai-clip==1.0.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
protobuf==5.29.5
psutil==7.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-build==0.18.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.19
tokenizers==0.15.2
tomli==2.2.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.37.0
triton==3.3.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
